# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Bruma API
  version: 1.0.0
  description: |
    MVP HTTP API (tenants, domains, templates, dispatch, webhooks, logs). Authenticate with `X-API-Key`
    (admin key or per-tenant key). Webhook endpoint uses HMAC signature headers instead of API keys.

    **Swagger UI:** `GET /docs` when `OPENAPI_DOCS_ENABLED` is true or outside production (see deployment config).
servers:
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /v1/tenants/{tenantId}/domains:
    parameters:
      - $ref: "#/components/parameters/TenantIdPath"
    get:
      tags: [Domains]
      summary: List domains for tenant
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Domain" }
    post:
      tags: [Domains]
      summary: Register domain (creates DKIM material)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fqdn]
              properties:
                fqdn: { type: string, example: mail.example.com }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
        "403": { description: Tenant inactive }
        "409": { description: Duplicate fqdn for tenant }
...
  /v1/tenants/{tenantId}/setup-checklist:
    get:
      tags: [Setup]
      summary: Onboarding checklist flags
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupChecklist" }
  /v1/tenants/{tenantId}/api-keys:
    parameters:
      - $ref: "#/components/parameters/TenantIdPath"
    get:
      tags: [ApiKeys]
      summary: List API key metadata (no secrets)
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ApiKeyMeta" }
    post:
      tags: [ApiKeys]
      summary: Issue API key (secret returned once)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 120 }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyCreated" }
  /v1/tenants/{tenantId}/api-keys/{keyId}:
    delete:
      tags: [ApiKeys]
      summary: Revoke API key
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - name: keyId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Revoked }
        "404": { description: Not found }
  /v1/tenants/{tenantId}/senders:
    parameters:
      - $ref: "#/components/parameters/TenantIdPath"
    get:
      tags: [Senders]
      summary: List senders
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Sender" }
    post:
      tags: [Senders]
      summary: Create sender (domain must be verified)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SenderCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sender" }
  /v1/tenants/{tenantId}/senders/{senderId}:
    get:
      tags: [Senders]
      summary: Get sender
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - name: senderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sender" }
    patch:
      tags: [Senders]
      summary: Update sender
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - name: senderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SenderPatch" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sender" }
  /v1/tenants/{tenantId}/branding:
    parameters:
      - $ref: "#/components/parameters/TenantIdPath"
    get:
      tags: [Branding]
      summary: Get branding profile
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branding" }
        "404": { description: Not configured }
    post:
      tags: [Branding]
      summary: Upsert branding profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BrandingBody" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branding" }
  /v1/tenants/{tenantId}/templates:
    parameters:
      - $ref: "#/components/parameters/TenantIdPath"
    get:
      tags: [Templates]
      summary: List templates (includes active_version_id when set)
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/TemplateSummary" }
    post:
      tags: [Templates]
      summary: Create template shell
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: [string, "null"] }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Template" }
  /v1/tenants/{tenantId}/templates/{templateId}/versions:
    get:
      tags: [Templates]
      summary: List template versions
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - $ref: "#/components/parameters/TemplateIdPath"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/TemplateVersion" }
    post:
      tags: [Templates]
      summary: Create template version (Handlebars body)
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - $ref: "#/components/parameters/TemplateIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body_html]
              properties:
                body_html: { type: string }
                variable_names:
                  type: array
                  items: { type: string }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateVersion" }
  /v1/tenants/{tenantId}/templates/{templateId}/publish:
    post:
      tags: [Templates]
      summary: Publish a version (single active)
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - $ref: "#/components/parameters/TemplateIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version_id]
              properties:
                version_id: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateVersion" }
  /v1/tenants/{tenantId}/templates/{templateId}/versions/{versionId}/preview:
    get:
      tags: [Templates]
      summary: Render preview HTML
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - $ref: "#/components/parameters/TemplateIdPath"
        - name: versionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: variables
          in: query
          required: false
          schema:
            type: string
            description: JSON object string, default "{}"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  template_id: { type: string, format: uuid }
                  version_id: { type: string, format: uuid }
                  html: { type: string }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    TenantIdPath:
      name: tenantId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    TemplateIdPath:
      name: templateId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    Domain:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        fqdn: { type: string }
        verification_status: { type: string }
        verification_error: { type: [string, "null"] }
        verified_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SetupChecklist:
      type: object
      properties:
        domain_registered: { type: boolean }
        domain_verified: { type: boolean }
        sender_created: { type: boolean }
        branding_configured: { type: boolean }
        template_published: { type: boolean }
        event_mapped: { type: boolean }
    ApiKeyMeta:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key_prefix: { type: string }
        revoked_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
    ApiKeyCreated:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key_prefix: { type: string }
        secret: { type: string, description: Shown once }
        created_at: { type: string, format: date-time }
    Sender:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        domain_id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        reply_to: { type: [string, "null"], format: email }
        state: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SenderCreate:
      type: object
      required: [domain_id, email, display_name]
      properties:
        domain_id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        reply_to: { type: [string, "null"], format: email }
    SenderPatch:
      type: object
      properties:
        display_name: { type: string }
        reply_to: { type: [string, "null"], format: email }
        state: { type: string, enum: [active, inactive] }
    Branding:
      type: object
      properties:
        tenant_id: { type: string, format: uuid }
        logo_url: { type: string, format: uri }
        color_primario: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
        updated_at: { type: string, format: date-time }
    BrandingBody:
      type: object
      required: [logo_url, color_primario]
      properties:
        logo_url: { type: string, format: uri }
        color_primario: { type: string }
    Template:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TemplateSummary:
      allOf:
        - { $ref: "#/components/schemas/Template" }
        - type: object
          properties:
            active_version_id: { type: [string, "null"], format: uuid }
    TemplateVersion:
      type: object
      properties:
        id: { type: string, format: uuid }
        template_id: { type: string, format: uuid }
        body_html: { type: string }
        variable_names:
          type: array
          items: { type: string }
        is_active: { type: boolean }
        version: { type: integer }
        created_at: { type: string, format: date-time }
...
security:
  - ApiKeyAuth: []